Remove Deprecated Assertion feature
- Remove deprecated old assertions capabilities
- Remove tests for the same
