+++ core/src/main/kotlin/io/specmatic/core/Resolver.kt
@@ -6,7 +6,6 @@ import io.specmatic.core.value.StringValue
 import io.specmatic.core.value.True
 import io.specmatic.core.value.Value
 import io.specmatic.test.ExampleProcessor
-import io.specmatic.test.asserts.WILDCARD_INDEX
 
 val actualMatch: (resolver: Resolver, factKey: String?, pattern: Pattern, sampleValue: Value) -> Result = { resolver: Resolver, factKey: String?, pattern: Pattern, sampleValue: Value ->
     resolver.actualPatternMatch(factKey, pattern, sampleValue)
