+++ core/src/main/kotlin/io/specmatic/core/ValidateUnexpectedKeys.kt
@@ -1,7 +1,6 @@
 package io.specmatic.core
 
 import io.specmatic.core.pattern.withoutOptionality
-import io.specmatic.test.asserts.ASSERT_KEYS
 
 object ValidateUnexpectedKeys: UnexpectedKeyCheck {
     override fun validate(pattern: Map<String, Any>, actual: Map<String, Any>): UnexpectedKeyError? {
@@ -16,7 +15,7 @@ object ValidateUnexpectedKeys: UnexpectedKeyCheck {
 
         val patternKeys = pattern.minus("...").keys.map { withoutOptionality(it) }
         val actualKeys = actual.keys.map { withoutOptionality(it) }
-        return actualKeys.minus(patternKeys.toSet().plus(ASSERT_KEYS)).map {
+        return actualKeys.minus(patternKeys.toSet()).map {
             UnexpectedKeyError(it)
         }
     }
@@ -24,6 +23,6 @@ object ValidateUnexpectedKeys: UnexpectedKeyCheck {
     override fun validateListCaseInsensitive(pattern: Map<String, Any>, actual: Map<String, Any>): List<UnexpectedKeyError> {
         val patternKeys = pattern.minus("...").keys.map { withoutOptionality(it).lowercase() }.toSet()
         val actualKeys = actual.keys.map { withoutOptionality(it) }
-        return actualKeys.filter { it.lowercase() !in patternKeys && it !in ASSERT_KEYS }.map { UnexpectedKeyError(it) }
+        return actualKeys.filter { it.lowercase() !in patternKeys }.map { UnexpectedKeyError(it) }
     }
 }
