+++ core/src/test/kotlin/integration_tests/LoadTestsFromExternalisedFiles.kt
@@ -1257,83 +1257,6 @@ class LoadTestsFromExternalisedFiles {
                 """.trimIndent())
             }
         }
-
-        @Test
-        fun `should be able to load partial example with missing discriminator but has asserts`() {
-            val specFile = File("src/test/resources/openapi/partial_with_discriminator/openapi.yaml")
-            val examplesDir = specFile.parentFile.resolve("example_with_asserts")
-
-            Flags.using(EXAMPLE_DIRECTORIES to examplesDir.canonicalPath) {
-                val feature = parseContractFileToFeature(specFile).copy(strictMode = true).loadExternalisedExamples()
-                feature.validateExamplesOrException()
-
-                var entity = JSONObjectValue()
-                val results = feature.executeTests(object: TestExecutor {
-                    override fun execute(request: HttpRequest): HttpResponse {
-                        return when(request.method) {
-                            "GET" -> HttpResponse(status = 200, body = JSONArrayValue(listOf(entity)))
-                            "POST" -> {
-                                assertThat(request.body).isEqualTo(parsedJSONObject("""{"petType": "cat", "color": "black"}"""))
-                                val withId = JSONObjectValue((request.body as JSONObjectValue).jsonObject.plus("id" to NumberValue(1)))
-                                HttpResponse(status = 201, body = withId).also { entity = withId }
-                            }
-                            else ->  throw Exception("Unknown method ${request.method}")
-                        }.also { println(listOf(request.toLogString(), it.toLogString()).joinToString(separator = "\n\n")) }
-                    }
-                }).results
-
-                assertThat(results).isNotEmpty.hasOnlyElementsOfType(Result.Success::class.java).hasSize(2)
-            }
-        }
-
-        @Test
-        fun `should complain when response doesn't match the asserts in the example`() {
-            val specFile = File("src/test/resources/openapi/partial_with_discriminator/openapi.yaml")
-            val examplesDir = specFile.parentFile.resolve("example_with_asserts")
-
-            Flags.using(EXAMPLE_DIRECTORIES to examplesDir.canonicalPath) {
-                val feature = parseContractFileToFeature(specFile).copy(strictMode = true).loadExternalisedExamples()
-                feature.validateExamplesOrException()
-
-                val invalidEntity = parsedJSONObject("""{ "id": 1, "color": "white", "petType": "dog" }""")
-                val results = feature.executeTests(object: TestExecutor {
-                    override fun execute(request: HttpRequest): HttpResponse {
-                        return when(request.method) {
-                            "GET" -> HttpResponse(status = 200, body = JSONArrayValue(listOf(invalidEntity)))
-                            "POST" -> {
-                                assertThat(request.body).isEqualTo(parsedJSONObject("""{"petType": "cat", "color": "black"}"""))
-                                val withId = JSONObjectValue((request.body as JSONObjectValue).jsonObject.plus("id" to NumberValue(1)))
-                                HttpResponse(status = 201, body = withId)
-                            }
-                            else ->  throw Exception("Unknown method ${request.method}")
-                        }.also { println(listOf(request.toLogString(), it.toLogString()).joinToString(separator = "\n\n")) }
-                    }
-                }).results
-
-                assertThat(results).hasSize(2)
-                assertThat(results.filterIsInstance<Result.Failure>()).hasSize(1)
-
-                val failure = results.filterIsInstance<Result.Failure>().first()
-                assertThat(failure.reportString()).containsIgnoringWhitespaces("""
-                In scenario "List all pets. Response: A list of pets"
-                API: GET /pets -> 200
-                ${
-                    toViolationReportString(
-                        breadCrumb = "RESPONSE.BODY[0].petType",
-                        details = "Expected \"dog\" to equal \"cat\"",
-                        StandardRuleViolation.VALUE_MISMATCH
-                    )
-                }
-                ${
-                    toViolationReportString(
-                        breadCrumb = "RESPONSE.BODY[0].color",
-                        details = "Expected \"white\" to equal \"black\"",
-                        StandardRuleViolation.VALUE_MISMATCH
-                    )
-                }
-                """.trimIndent())
-            }
-        }
     }
 
     @Nested
