+++ core/src/test/kotlin/io/specmatic/core/FeatureTest.kt
@@ -3728,136 +3728,6 @@ paths:
         }
     }
 
-    @Test
-    fun `should perform assertions on the response for test created from example file`(@TempDir tempDir: File) {
-        val apiSpecification = """
-        openapi: 3.0.0
-        info:
-          title: Simple API
-          version: 1.0.0
-        paths:
-          /test:
-            post:
-              summary: Test Example
-              requestBody:
-                required: true
-                content:
-                  application/json:
-                    schema:
-                      ${"$"}ref: '#/components/schemas/ExampleRequest'
-              responses:
-                '200':
-                  description: Successful response
-                  content:
-                    application/json:
-                      schema:
-                        ${"$"}ref: '#/components/schemas/ExampleResponse'
-        components:
-          schemas:
-            ExampleRequest:
-              type: object
-              required:
-                - name
-                - age
-              properties:
-                name:
-                  type: string
-                age:
-                  type: integer
-                isEligible:
-                  type: boolean
-            ExampleResponse:
-              allOf:
-                - ${"$"}ref: '#/components/schemas/ExampleRequest'
-                - type: object
-                  required:
-                    - id
-                  properties:
-                    id:
-                      type: string
-        """.trimIndent()
-        val example = """
-        {
-          "partial": {
-            "http-request": {
-              "method": "POST",
-              "path": "/test",
-              "body": {
-                "name": "(string)",
-                "isEligible": true
-              }
-            },
-            "http-response": {
-              "status": 200,
-              "body": {
-                "id": "(string)",
-                "name": "${"$"}eq(REQUEST.BODY.name)",
-                "age": "${"$"}eq(REQUEST.BODY.age)",
-                "isEligible": "${"$"}eq(REQUEST.BODY.isEligible)"
-              }
-            }
-          }
-        }
-        """.trimIndent()
-        val dictionary = """
-        ExampleRequest:
-          name: John Doe
-          age: 999
-          isEligible: false
-        """.trimIndent()
-
-        val apiSpecFile = tempDir.resolve("api.yaml").apply { writeText(apiSpecification) }
-        val examplesDir = tempDir.resolve("api_examples").apply { mkdirs() }
-        val exampleFile = examplesDir.resolve("example.json").apply { writeText(example) }
-        tempDir.resolve("api_dictionary.yaml").apply { writeText(dictionary) }
-
-        val feature = parseContractFileToFeature(apiSpecFile)
-        val contractTest = feature.createContractTestFromExampleFile(exampleFile.canonicalPath).value
-
-        val expectedRequestBody = parsedJSONObject("""{"name" : "John Doe", "isEligible" : true, "age" : 999}""")
-        val result = contractTest.runTest(object : TestExecutor {
-            override fun execute(request: HttpRequest): HttpResponse {
-                assertThat(request.body).isEqualTo(expectedRequestBody)
-                return HttpResponse.ok(
-                    parsedJSONObject("""{"name" : "Jane Doe", "isEligible" : false, "age" : 123, "id": "10"}""")
-                ).also {
-                    println(request.toLogString())
-                    println()
-                    println(it.toLogString())
-                }
-            }
-        }).result
-
-        assertThat(result).isInstanceOf(Result.Failure::class.java)
-        assertThat(result.reportString()).isEqualToNormalizingWhitespace(
-            """
-        In scenario "Test Example. Response: Successful response"
-        API: POST /test -> 200
-        ${
-                toViolationReportString(
-                    breadCrumb = "RESPONSE.BODY.name",
-                    details = "Expected \"Jane Doe\" to equal \"John Doe\"",
-                    StandardRuleViolation.VALUE_MISMATCH
-                )
-            }
-        ${
-                toViolationReportString(
-                    breadCrumb = "RESPONSE.BODY.age",
-                    details = "Expected 123 to equal 999",
-                    StandardRuleViolation.VALUE_MISMATCH
-                )
-            }
-        ${
-                toViolationReportString(
-                    breadCrumb = "RESPONSE.BODY.isEligible",
-                    details = "Expected false to equal true",
-                    StandardRuleViolation.VALUE_MISMATCH
-                )
-            }
-        """.trimIndent()
-        )
-    }
-
     @Nested
     inner class GenerateDiscriminatorDetailsBasedRequestResponsePairsTest {
         private val feature = Feature(name = "feature", protocol = SpecmaticProtocol.HTTP)
