+++ core/src/test/kotlin/io/specmatic/core/ResolverKtTest.kt
@@ -4,7 +4,6 @@ import io.specmatic.core.pattern.AnyValuePattern
 import org.assertj.core.api.Assertions.*
 import org.junit.jupiter.api.Test
 import io.specmatic.core.pattern.ContractException
-import io.specmatic.test.asserts.WILDCARD_INDEX
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.Arguments
 import org.junit.jupiter.params.provider.MethodSource
