+++ core/src/test/kotlin/io/specmatic/test/ExampleProcessorTest.kt
@@ -12,7 +12,7 @@ import io.specmatic.core.value.JSONArrayValue
 import io.specmatic.core.value.JSONObjectValue
 import io.specmatic.core.value.NumberValue
 import io.specmatic.core.value.StringValue
-import io.specmatic.test.asserts.AssertComparisonTest.Companion.toFactStore
+import io.specmatic.test.ExampleProcessor.toFactStore
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.AfterEach
 import org.junit.jupiter.api.BeforeEach
