--- core/src/test/resources/openapi/partial_with_discriminator/example_with_asserts/get_example.json
@@ -1,29 +1,0 @@
-{
-  "partial": {
-    "http-request": {
-      "path": "/pets",
-      "method": "GET"
-    },
-    "http-response": {
-      "status": 200,
-      "body": [
-        {
-          "id": "$array_has(ENTITY.id)",
-          "$if": {
-            "$conditions": {
-              "id": "$eq(ENTITY.id)"
-            },
-            "$then": {
-              "petType": "$eq(ENTITY.petType)",
-              "color": "$eq(ENTITY.color)"
-            }
-          }
-        }
-      ],
-      "status-text": "Created",
-      "headers": {
-        "Content-Type": "application/json"
-      }
-    }
-  }
-}
