--- core/src/test/resources/openapi/partial_with_discriminator/example_with_asserts/post_example.json
@@ -1,21 +1,0 @@
-{
-  "partial": {
-    "http-request": {
-      "path": "/pets",
-      "method": "POST",
-      "body": "(Cat)"
-    },
-    "http-response": {
-      "status": 201,
-      "body": {
-        "id": "(number)",
-        "petType": "cat",
-        "color": "$eq(REQUEST.BODY.color)"
-      },
-      "status-text": "Created",
-      "headers": {
-        "Content-Type": "application/json"
-      }
-    }
-  }
-}
